moved debug view to the separate file
